fix(merge): unify loadState — family seq convention (gates before allocation), 8-arg constructor with progress, fixture builders aligned, mac test updated to family semantics

+++ crates/opentray-ext-webview/src/macos/load_state.rs
@@ -244,7 +244,7 @@ impl LoadStateNavigationDelegate {
             &ivars.owner.window_id,
             WebviewLoadPhase::Failed,
             url,
-            Some(error_code as i64),
+            Some(error_code as i32),
             None,
         );
         super::push_event_frame(&ivars.outbox, frame);
